+++ test/run_tests.bash
@@ -43,3 +43,4 @@ do
    $p ./shadow_package/shadow_package.py
 done
 
+echo